[FogWorldMap] Temporary fix BP_WorldMapPawn
Bug fixed: when the user switches from the world map to the boat and that he was moving or rotating, the "released" event is not triggered. Thus, when the user comes back to the world map, the gate is still opened and the movement is triggered without his will. This is because the gates can't close when the switch is done because the "released" events never happen.